readme

+++ app/src/themes.ts
@@ -12,8 +12,8 @@ const sheetIndex = 1;
 export function setTheme(character) {
   let sheet = document.styleSheets[sheetIndex];
 
-  // @ts-ignore
-  if (sheet.cssRules[0].selectorText === ':root') {
+  //@ts-ignore
+  if ((sheet.cssRules[0] !== undefined) && (sheet.cssRules[0].selectorText === ':root')) {
     sheet.deleteRule(0);
   }
   sheet.insertRule(getTheme(character), 0);
